refactor(mode): move default slice under src/mode/default; drop app_mode.rs
default_input.rs/default_webview.rs -> src/mode/default/{input,webview}.rs; the
Default-mode UI lifecycle moves from app_mode.rs into src/mode/default.rs. main.rs
registers the same plugins via crate::mode::default. No behavior change.
