+++ src/mode/default.rs
@@ -1,4 +1,12 @@
-//! Default-mode UI subtree lifecycle plugin.
+//! Default-mode (`AppMode::Default`) processing: the single-PTY shell UI
+//! lifecycle, host-side input gates/shortcuts, and webview pointer routing.
+
+mod input;
+mod webview;
+
+pub(crate) use input::DefaultHostInputPlugin;
+pub(crate) use input::should_disable_input;
+pub(crate) use webview::DefaultWebviewPointerPlugin;
 
 use crate::mode::AppMode;
 use crate::ui::UiRoot;
